#27 : Event driven effects (hot/dot/expiration)

+++ MysteryMud.ConsoleApp3/Components/Effects/DamageOverTime.cs
@@ -6,6 +6,6 @@ struct DamageOverTime
 {
     public int Damage;
     public DamageType DamageType;
-    public int TickRate;
-    public int NextTick;
+    public long TickRate; // How many ticks between each damage application
+    public long NextTick; // The tick at which the next damage application will occur
 }
