+++ MysteryMud.ConsoleApp3/Components/Effects/HealOverTime.cs
@@ -3,6 +3,6 @@
 struct HealOverTime
 {
     public int Heal;
-    public int TickRate;
-    public int NextTick;
+    public long TickRate; // How many ticks between each heal
+    public long NextTick; // The tick at which the next heal will occur
 }
